lab10/inventoryservice_ps/requirements.txt
flask
google-cloud-pubsub==2.4.1